Abstract out loadAudio method
